+++ packages/terra-site/CHANGELOG.md
@@ -10,6 +10,7 @@ Unreleased
 ------------------
 ### Changed
 * Minor version bump
+* Updated terra-toggle examples
 
 1.12.0 - (September 26, 2017)
 ------------------