+++ packages/terra-site/CHANGELOG.md
@@ -3,8 +3,12 @@ Changelog
 
 Unreleased
 ----------
-### Fixed
-* Typos
+
+1.13.0 - (October 6, 2017)
+------------------
+### Changed
+* Minor version bump
+* Updated terra-toggle examples
 
 1.12.0 - (September 26, 2017)
 ------------------